created a random sample of data points on a route

+++ ENG4810_Software/ENG4810_Software/DataSample.cs
@@ -9,16 +9,16 @@ namespace ENG4810_Software
     class DataSample
     {
         public string time;                // string representation of the time in ISO 8061 format
-        public float latitude;             // latitude coordinates         units: n/a
-        public float longitude;            // longitude coordinates        units: n/a
-        public float temperature;          // temperature reading          units: C
-        public float uv;                   // uv light intensity reading   units: mW/cm^2
+        public double latitude;             // latitude coordinates         units: n/a
+        public double longitude;            // longitude coordinates        units: n/a
+        public double temperature;          // temperature reading          units: C
+        public int uv;                   // uv light intensity reading   units: mW/cm^2
         public int pressure;               // absolute pressure reading    units: kPa
         public int sound;                  // SPL A weighted sound level   units: dBA
         public int luminosity;             // Luminosity in reading in     units: Lux
         public int humidity;               // humidity reading             units: %RH
-        public List<float> acceleration;   // XYZ acceleration readings    Units: m/s^2
-        public List<float> magnetic_Field; // XYZ magnetic field readings  units: uT
+        public List<double> acceleration;   // XYZ acceleration readings    Units: m/s^2
+        public List<int> magnetic_Field; // XYZ magnetic field readings  units: uT
         /*
          *Constructor initalises all variable to empty/0 
          */
@@ -34,8 +34,8 @@ namespace ENG4810_Software
             this.sound = 0;
             this.luminosity = 0;
             this.humidity = 0;
-            this.acceleration = new List<float>();
-            this.magnetic_Field = new List<float>();
+            this.acceleration = new List<double>();
+            this.magnetic_Field = new List<int>();
             this.temperature = 0;
         }
     }
